moved mask_email to utils.py
