+++ lastuser_oauth/views/helpers.py
@@ -279,19 +279,3 @@ def set_loginmethod_cookie(response, value):
         expires=datetime.utcnow() + timedelta(days=365),   # Expire one year from now
         httponly=True)
     return response
-
-
-def mask_email(email):
-    """
-    Masks an email address
-
-    >>> mask_email(u'foobar@example.com')
-    u'f*****@e**********'
-
-    """
-    if '@' not in email:
-        return email
-    username, domain = email.split('@')
-    masked_username = username.replace(username[1:], '*' * len(username[1:]))
-    masked_domain = domain.replace(domain[1:], '*' * len(domain[1:]))
-    return u'@'.join([masked_username, masked_domain])
